Changing the 'find' functionality
Now find is done by means of a RegExp match, both when
searching among the the Mod:Fun calls, as well as when
searching among the argument values.

Searching among the return values are now done by prefixing
the search string with: ~r

Example:

  tlist> f ~rsome string

Also, updating the docs to fit hex.pm.